# perf: Avoid hashing together default digests (zeros) in MT construction