When matrices have #rows != pow-of-two, there is no reason we should be padding to the next power of two and hashing all the zeros together. We can save prover time by avoiding hashing the trailing zeros.

These changes affect the root computation (the digest changes).
The API for opening & verification stay the same.
The changes make the code a bit less elegant.
Also note that if higher arity trees are ever introduced, the `_padded` lengths would need to change from even, to divisible by the arity (ariteven?).

The change is best illustrated with these two matrices: this exact test is included, see commit 62d02bcc373c016f2ec716fd98e9ac69b8cf6cfd which adds the test case for current construction (initially broken) and a59d70663c0e0580d79f5a3731785d7984b6e7e3 which fixes it to adapt to this PR.
mat1: 5 rows
mat2: 3 rows

**Previous MT construction**
```
let expected_result = compress.compress([
    compress.compress([
        compress.compress([
            compress.compress([mat_1_leaf_hashes[0], mat_1_leaf_hashes[1]]),
            mat_2_leaf_hashes[0],
        ]),
        compress.compress([
            compress.compress([mat_1_leaf_hashes[2], mat_1_leaf_hashes[3]]),
            mat_2_leaf_hashes[1],
        ]),
    ]),
    compress.compress([
        compress.compress([
            compress.compress([mat_1_leaf_hashes[4], default_digest]),
            mat_2_leaf_hashes[2],
        ]),
        compress.compress([
            compress.compress([default_digest, default_digest]),
            default_digest,
        ]),
    ]),
]);
```

**Improved MT construction**
```
let expected_result = compress.compress([
    compress.compress([
        compress.compress([
            compress.compress([mat_1_leaf_hashes[0], mat_1_leaf_hashes[1]]),
            mat_2_leaf_hashes[0],
        ]),
        compress.compress([
            compress.compress([mat_1_leaf_hashes[2], mat_1_leaf_hashes[3]]),
            mat_2_leaf_hashes[1],
        ]),
    ]),
    compress.compress([
        compress.compress([
            compress.compress([mat_1_leaf_hashes[4], default_digest]),
            mat_2_leaf_hashes[2],
        ]),
        default_digest,
    ]),
]);
```


# Bench results
All single threaded

### current, Blake3
```
MerkleTreeMmcs::<p3_symmetric::serializing_hasher::SerializingHasher32<p3_blake3::Blake3>, p3_symmet...
                        time:   [128.64 ms 128.73 ms 128.81 ms]
```

### improved, Blake3
```
MerkleTreeMmcs::<p3_symmetric::serializing_hasher::SerializingHasher32<p3_blake3::Blake3>, p3_symmet...
                        time:   [119.55 ms 119.87 ms 120.17 ms]
                        change: [-7.1368% -6.9825% -6.8059%] (p = 0.00 < 0.05)
                        Performance has improved.
```

### current, Poseidon2
```
MerkleTreeMmcs::<p3_symmetric::sponge::PaddingFreeSponge<p3_poseidon2::Poseidon2<p3_monty_31::monty_...
                        time:   [351.35 ms 351.40 ms 351.45 ms]
```

### improved, Poseidon2
```
MerkleTreeMmcs::<p3_symmetric::sponge::PaddingFreeSponge<p3_poseidon2::Poseidon2<p3_monty_31::monty_...
                        time:   [317.10 ms 317.15 ms 317.19 ms]
                        change: [-9.7650% -9.7472% -9.7289%] (p = 0.00 < 0.05)
                        Performance has improved.
```

### Parallel
-29.180% for Blake3
-45.842% for Poseidon2

https://github.com/NP-Eng/Plonky3/tree/mmagician/hash-zeroes-par/
vs. 
https://github.com/NP-Eng/Plonky3/tree/mmagician/skip-hashing-zeroes-par