@mmagician 2024-10-28
> Only suggestion is that the padding might be a bit easier to read as `next_multiple_of(2)`? It could potentially be slower due to division, but it's `#[inline]` so I'd expect that to get optimized out. It's minor though, we can also merge as-is and tweak it later.

Sure, let's merge as-is then, and I can later bench whether the proposed modification has any impact on the performance